Energy deviation, simulation speed, legend on the plot, yaml fix
Fixed incorrect loading of yaml, occurring when there are only integers in the file

+++ celestial_body_object.py
@@ -20,7 +20,7 @@ class CelestialBodyObject:
     body: CelestialBody
     name: str
     color: (int, int, int)
-    size: int
+    size: float
     trajectory: CelestialBodyTrajectory = None
